Record minimum number of frames after detection
Records a minimum number of frames after every detection event.
Allows capture of what are most likely false negatives for analysis
and creates an image series with better continuity with the advantage
that we can also see the behaviour of the control system in response
to the stream of detections and false negatives.