Add script for "up/down" cam experiment
Imagining an scenario where we want to do something like
detect_track_zoom when an animal might go well above (or below if
camera pointed up) the current horizon plane of a pinpoint setup,
trying an arrangement where there is a camera pointed up couple with
one pointed down.  An example would be with an animal that climbs a
tree trunk that is much higher than the plan with situate the camera
on but when the dominant animal activity warrants a downward-facing
camera.