Save unboxed frame even if DRAW_BOX is on
Before if DRAW_BOX was set to True, the frames saved to disk (if
frames are set to save to disk) would have boxes drawn on them. This
is non-ideal as it is much better use post-processing utilities to add
boxes from lboxes data to raw frames than to have no clean frame data.