Move early experiments to aptly named folder
Cleaning up clutter in main area of repo by moving these early
experiments into their own folder.  Didn't want to nuke them outright
as not sure there aren't some things in there that might bear
revisiting at a later stage and, even as they are in the history,
having a folder sitting there with them in it will serve as a reminder
that (some day) that might be worth looking into.