Create generic ABC from MotorController
And make a derived class that implements the current main "calm"
controller that has been the recent mainstay.  This is in preparation
to make controllers that work differently and towards different goals,
particularly one that implements something closer to the "twitchier"
control of before that prioritized quick response over worrying about
the "weaving around" artifact when locked in on a target that isn't
doing much moving around.