Don't create display handler when in headless mode
Before this commit, code was creating a display handler and thus also
a camera stream capture object even if in headless mode.  Mechanics
within the class were duly not trying to display to a screen when
headless but was a waste of resources to even generate the stream
capture connection.