Simplify image timestamp preparation
Use only one call to datetime.now instead of nutty combination of
datetime, time, and math. Possibly solve an existing bug (occasional
images timestamped out of order) that perhaps is being caused by
getting the microseconds from datetime.now immediately after getting
the rest of the time from time.  Even if this isn't the source of that
image-out-of-order bug that has been cropping up, it certainly seems
like a bug waiting to happen.