Return 0 for errors when no target_lbox
Our primary goal here is to be able to calculate errors every control
loop in an application and not just the ones where there is a
detection.  Before this in the detect_track_zoom application
calc_errors was only being called if a target_lbox was detected.

Our proximal goal was/is to be able to have zoom control take into
account whether there is a target cycle-to-cycle and to further
encapsulate zoom logic away from the application and into the
MotorController.  However, it is probably also true that it is better
for pan and tilt errors to be evaluated every cycle, too, and not just
ones in which there was detection.